+++ src/prefect/utilities/__init__.py
@@ -1,7 +1,7 @@
 # Licensed under LICENSE.md; also available at https://www.prefect.io/licenses/alpha-eula
 
 import prefect.utilities.logging
-
+import prefect.utilities.cli
 import prefect.utilities.collections
 import prefect.utilities.datetimes
 import prefect.utilities.exceptions
